Implement useAuthEmulator and remove auth.useEmulator.
