+++ packages-exp/auth-exp/src/core/auth/auth_impl.test.ts
@@ -461,126 +461,3 @@ describe('core/auth/auth_impl', () => {
     });
   });
 });
-
-// These tests are separate because they are using a different auth with
-// separate setup and config
-describe('core/auth/auth_impl useEmulator', () => {
-  let auth: TestAuth;
-  let user: User;
-  let normalEndpoint: fetch.Route;
-  let emulatorEndpoint: fetch.Route;
-
-  beforeEach(async () => {
-    auth = await testAuth();
-    user = testUser(_castAuth(auth), 'uid', 'email', true);
-    fetch.setUp();
-    normalEndpoint = mockEndpoint(Endpoint.DELETE_ACCOUNT, {});
-    emulatorEndpoint = fetch.mock(
-      `http://localhost:2020/${endpointUrl(Endpoint.DELETE_ACCOUNT).replace(
-        /^.*:\/\//,
-        ''
-      )}`,
-      {}
-    );
-  });
-
-  afterEach(() => {
-    fetch.tearDown();
-    sinon.restore();
-
-    // The DOM persists through tests; remove the banner if it is attached
-    const banner =
-      typeof document !== 'undefined'
-        ? document.querySelector('.firebase-emulator-warning')
-        : null;
-    if (banner) {
-      banner.parentElement?.removeChild(banner);
-    }
-  });
-
-  context('useEmulator', () => {
-    it('fails if a network request has already been made', async () => {
-      await user.delete();
-      expect(() => auth.useEmulator('http://localhost:2020')).to.throw(
-        FirebaseError,
-        'auth/emulator-config-failed'
-      );
-    });
-
-    it('updates the endpoint appropriately', async () => {
-      auth.useEmulator('http://localhost:2020');
-      await user.delete();
-      expect(normalEndpoint.calls.length).to.eq(0);
-      expect(emulatorEndpoint.calls.length).to.eq(1);
-    });
-
-    it('checks the scheme properly', () => {
-      expect(() => auth.useEmulator('http://localhost:2020')).not.to.throw;
-      delete auth.config.emulator;
-      expect(() => auth.useEmulator('https://localhost:2020')).not.to.throw;
-      delete auth.config.emulator;
-      expect(() => auth.useEmulator('ssh://localhost:2020')).to.throw(
-        FirebaseError,
-        'auth/invalid-emulator-scheme'
-      );
-      delete auth.config.emulator;
-      expect(() => auth.useEmulator('localhost:2020')).to.throw(
-        FirebaseError,
-        'auth/invalid-emulator-scheme'
-      );
-    });
-
-    it('attaches a banner to the DOM', () => {
-      auth.useEmulator('http://localhost:2020');
-      if (typeof document !== 'undefined') {
-        const el = document.querySelector('.firebase-emulator-warning')!;
-        expect(el).not.to.be.null;
-        expect(el.textContent).to.eq(
-          'Running in emulator mode. ' +
-            'Do not use with production credentials.'
-        );
-      }
-    });
-
-    it('logs out a warning to the console', () => {
-      sinon.stub(console, 'info');
-      auth.useEmulator('http://localhost:2020');
-      expect(console.info).to.have.been.calledWith(
-        'WARNING: You are using the Auth Emulator,' +
-          ' which is intended for local testing only.  Do not use with' +
-          ' production credentials.'
-      );
-    });
-
-    it('logs out the warning but has no banner if disableBanner true', () => {
-      sinon.stub(console, 'info');
-      auth.useEmulator('http://localhost:2020', { disableWarnings: true });
-      expect(console.info).to.have.been.calledWith(
-        'WARNING: You are using the Auth Emulator,' +
-          ' which is intended for local testing only.  Do not use with' +
-          ' production credentials.'
-      );
-      if (typeof document !== 'undefined') {
-        expect(document.querySelector('.firebase-emulator-warning')).to.be.null;
-      }
-    });
-  });
-
-  context('toJSON', () => {
-    it('works when theres no current user', () => {
-      expect(JSON.stringify(auth)).to.eq(
-        '{"apiKey":"test-api-key","authDomain":"localhost","appName":"test-app"}'
-      );
-    });
-
-    it('also stringifies the current user', () => {
-      auth.currentUser = ({
-        toJSON: (): object => ({ foo: 'bar' })
-      } as unknown) as User;
-      expect(JSON.stringify(auth)).to.eq(
-        '{"apiKey":"test-api-key","authDomain":"localhost",' +
-          '"appName":"test-app","currentUser":{"foo":"bar"}}'
-      );
-    });
-  });
-});
