+++ packages-exp/auth-exp/src/model/auth.ts
@@ -56,6 +56,7 @@ export interface Auth extends externs.Auth {
   _startProactiveRefresh(): void;
   _stopProactiveRefresh(): void;
   _getPersistence(): string;
+  _useEmulator(url: string, options?: { disableWarnings: boolean }): void;
 
   readonly name: AppName;
   readonly config: ConfigInternal;
